moved some methods into network_operations and updated everything accodringly
