+++ tensornetwork/network_components.py
@@ -1094,6 +1094,468 @@ class Edge:
     return self.disconnect('__unnamed_edge__', '__unnamed_dge__')
 
 
+def get_shared_edges(node1: BaseNode, node2: BaseNode) -> Set[Edge]:
+  """Get all edges shared between two nodes.
+
+  Args:
+    node1: The first node.
+    node2: The second node.
+
+  Returns:
+    A (possibly empty) `set` of `Edge`s shared by the nodes.
+  """
+  nodes = {node1, node2}
+  shared_edges = set()
+  # Assuming the network is well formed, all of the edges shared by
+  # these two nodes will be stored in just one of the nodes, so we only
+  # have to do this loop once.
+  for edge in node1.edges:
+    if set(edge.get_nodes()) == nodes:
+      shared_edges.add(edge)
+  return shared_edges
+
+
+def get_parallel_edges(edge: Edge) -> Set[Edge]:
+  """
+  Get all of the edges parallel to the given `edge`.
+  Args:
+    edge: The given edge.
+
+  Returns:
+    A `set` of all of the edges parallel to the given edge 
+    (including the given edge).
+  """
+  return get_shared_edges(edge.node1, edge.node2)
+
+
+def get_all_nondangling(
+    nodes: Union[List[BaseNode], Set[BaseNode]]) -> Set[Edge]:
+  """Return the set of all non-dangling edges."""
+  edges = set()
+  for node in nodes:
+    edges |= node.get_all_nondangling()
+  return edges
+
+
+def _flatten_trace_edges(edges: List[Edge],
+                         new_edge_name: Optional[Text] = None) -> Edge:
+  """Flatten trace edges into single edge.
+
+  Args:
+    edges: List of trace edges to flatten
+    new_edge_name: Optional name of the new edge created.
+
+  Returns:
+    The new edge that represents the flattening of the given edges.
+  """
+  node = edges[0].node1  # We are in the trace case, so this is the only node.
+  backend = node.backend
+  # Flatten all of the edge's axes into a a single list.
+  perm_back = [min(e.axis1, e.axis2) for e in edges]
+  perm_back += [max(e.axis1, e.axis2) for e in edges]
+  perm_front = set(range(len(node.edges))) - set(perm_back)
+  perm_front = sorted(perm_front)
+  perm = perm_front + perm_back
+  new_dim = backend.prod([backend.shape(node.tensor)[e.axis1] for e in edges])
+  node.reorder_axes(perm)
+  unaffected_shape = backend.shape(node.tensor)[:len(perm_front)]
+  new_shape = backend.concat([unaffected_shape, [new_dim, new_dim]], axis=-1)
+  node.tensor = backend.reshape(node.tensor, new_shape)
+  edge1 = Edge(node1=node, axis1=len(perm_front), name="TraceFront")
+  edge2 = Edge(node1=node, axis1=len(perm_front) + 1, name="TraceBack")
+  node.edges = node.edges[:len(perm_front)] + [edge1, edge2]
+  new_edge = connect(edge1, edge2, new_edge_name)
+  # pylint: disable=expression-not-assigned
+  [edge.disable() for edge in edges]  #disable edges!
+  return new_edge
+
+
+def flatten_edges(edges: List[Edge],
+                  new_edge_name: Optional[Text] = None) -> Edge:
+  """Flatten edges into single edge.
+
+  If two nodes have multiple edges connecting them, it may be
+  beneficial to flatten these edges into a single edge to avoid having several
+  unnecessary trace edges. This can speed up computation time and reduce
+  memory cost.
+
+  Warning: This will remove all axes names.
+
+  Args:
+    edges: A list of edges to flatten.
+    backend: A backend object
+    new_edge_name: Optional name to give to the newly created edge.
+
+  Returns:
+    The new flattened edge.
+
+  Raises:
+    ValueError: If edges is an empty list.
+    ValueError: If not all of the edges connect to the same node(s).
+    ValueError: If one of the nodes connecting to these edges does not have
+      edge definitions for all of its axes.
+  """
+  if not edges:
+    raise ValueError("At least 1 edge must be given.")
+
+  backends = [edge.node1.backend for edge in edges] + [
+      edge.node2.backend for edge in edges if edge.node2 is not None
+  ]
+
+  if not all([b.name == backends[0].name for b in backends]):
+    raise ValueError("Not all backends are the same.")
+  backend = backends[0]
+  if len(edges) == 1:
+    return edges[0]  # Don't bother with reshaping.
+  # Set equality is transitive (a=b, b=c, therefore a=c) so it is only
+  # necessary to compare the first edge against the rest.
+  expected_nodes = set(edges[0].get_nodes())
+  for edge in edges:
+    if expected_nodes != set(edge.get_nodes()):
+      raise ValueError(
+          "Two edges do not share the same nodes. "
+          "'{}'s nodes: '{}', '{}'. '{}'s nodes: '{}', '{}'".format(
+              edges[0], edges[0].node1, edges[0].node2, edge, edge.node1,
+              edge.node2))
+  if len(expected_nodes) == 1:
+    return _flatten_trace_edges(edges, new_edge_name)  #disables edges
+  # Flatten standard or dangling edges.
+  new_dangling_edges = []
+  for node in expected_nodes:
+    # Required for dangling case.
+    if node is None:
+      continue
+    axis_names = node.axis_names
+    perm_back = []
+    for edge in edges:
+      # There will only be 1 edge since we are in the standard edge case.
+      perm_back.append(node.edges.index(edge))
+    perm_front = sorted(set(range(len(node.edges))) - set(perm_back))
+    node.reorder_axes(perm_front + perm_back)
+    old_tensor_shape = backend.shape(node.tensor)
+    # Calculate the new axis dimension as a product of the other
+    # axes dimensions.
+    flattened_axis_dim = backend.prod(old_tensor_shape[len(perm_front):])
+    new_tensor_shape = backend.concat(
+        [old_tensor_shape[:len(perm_front)], [flattened_axis_dim]], axis=-1)
+    new_tensor = backend.reshape(node.tensor, new_tensor_shape)
+    # Modify the node in place. Currently, this is they only method that
+    # modifies a node's tensor.
+    node.tensor = new_tensor
+    # This Edge is required for the connect call later.
+    edge = Edge(node1=node, axis1=len(perm_front), name=new_edge_name)
+    # Do not set the signature of 'edge' since it is dangling.
+    node.edges = node.edges[:len(perm_front)] + [edge]
+    new_dangling_edges.append(edge)
+    # TODO: Allow renaming of the new axis.
+    if axis_names:
+      node.axis_names = [axis_names[n] for n in range(len(node.edges))]
+    else:
+      node.axis_names = None
+
+  node1, node2 = tuple(expected_nodes)
+  # Sets are returned in a random order, so this is how we deal with
+  # dangling edges.
+  # pylint: disable=expression-not-assigned
+  [edge.disable() for edge in edges]  #disable edges!
+  if node1 is None or node2 is None:
+    return new_dangling_edges[0]
+
+  return connect(new_dangling_edges[0], new_dangling_edges[1], new_edge_name)
+
+
+def flatten_edges_between(
+    node1: BaseNode,
+    node2: BaseNode,
+) -> Optional[Edge]:
+  """Flatten all of the edges between the given two nodes.
+
+  Args:
+    node1: The first node.
+    node2: The second node.
+
+  Returns:
+    The flattened `Edge` object. If there was only one edge between the two
+      nodes, then the original edge is returned. If there where no edges
+      between the nodes, a None is returned.
+  """
+  shared_edges = get_shared_edges(node1, node2)
+  if shared_edges:
+    return flatten_edges(list(shared_edges))
+  return None
+
+
+def flatten_all_edges(nodes: List[BaseNode]) -> List[Edge]:
+  """Flatten all edges in the network.
+
+  Returns:
+    A list of all the flattened edges. If there was only one edge between
+    two given nodes, that original edge is included in this list.
+  """
+  flattened_edges = []
+  for edge in get_all_nondangling(nodes):
+    if not edge.is_disabled:
+      flat_edge = flatten_edges_between(edge.node1, edge.node2)
+      flattened_edges.append(flat_edge)
+  return flattened_edges
+
+
+def _remove_trace_edge(edge: Edge, new_node: BaseNode) -> None:
+  """Collapse a trace edge. `edge` is disabled before returning.
+
+  Take a trace edge (i.e. with edge.node1 = edge.node2),
+  remove it, update the axis numbers of all remaining edges
+  and move them to `new_node`.
+
+  Args:
+    edge: The edge to contract.
+    new_node: The new node created after contraction.
+
+  Returns:
+    None
+
+  Raises:
+    ValueError: If edge is not a trace edge.
+  """
+  if edge.is_dangling():
+    raise ValueError("Attempted to remove dangling edge '{}'.".format(edge))
+  if edge.node1 is not edge.node2:
+    raise ValueError("Edge '{}' is not a trace edge.".format(edge))
+  axes = sorted([edge.axis1, edge.axis2])
+  node_edges = edge.node1.edges[:]
+  node_edges.pop(axes[0])
+  node_edges.pop(axes[1] - 1)
+  seen_edges = set()
+  for tmp_edge in node_edges:
+    if tmp_edge in seen_edges:
+      continue
+    else:
+      seen_edges.add(tmp_edge)
+    if tmp_edge.node1 is edge.node1:
+      to_reduce = 0
+      to_reduce += 1 if tmp_edge.axis1 > axes[0] else 0
+      to_reduce += 1 if tmp_edge.axis1 > axes[1] else 0
+      tmp_edge.axis1 -= to_reduce
+      tmp_edge.node1 = new_node
+    if tmp_edge.node2 is edge.node1:
+      to_reduce = 0
+      to_reduce += 1 if tmp_edge.axis2 > axes[0] else 0
+      to_reduce += 1 if tmp_edge.axis2 > axes[1] else 0
+      tmp_edge.axis2 -= to_reduce
+      tmp_edge.node2 = new_node
+  # Update edges for the new node.
+  for i, e in enumerate(node_edges):
+    new_node.add_edge(e, i)
+  edge.node1.fresh_edges(edge.node1.axis_names)
+  edge.disable()  #disabled edge!
+
+
+def _remove_edges(edges: Set[Edge], node1: BaseNode, node2: BaseNode,
+                  new_node: BaseNode) -> None:
+  """
+
+  Takes a set of `edges` shared between `node1` and `node2` to be contracted
+  over, and moves all other uncontracted edges from `node1` and `node2` to
+  `new_node`.
+  The nodes that currently share the edges in `edges` must be supplied as
+  `node1` and `node2`. The ordering of `node1` and `node2` must match the
+  axis ordering of `new_node` (as determined by the contraction procedure).
+  `node1` and `node2` get both a fresh set edges.
+  `edges` are disabled before returning.
+  Args:
+    edges: The edges to contract.
+    node1: The old node that supplies the first edges of `new_node`.
+    node2: The old node that supplies the last edges of `new_node`.
+    new_node: The new node that represents the contraction of the two old
+      nodes.
+  Returns:
+    node1, node2L
+  Raises:
+    Value Error: If edge isn't in the network.
+  """
+  if node1 is node2:
+    raise ValueError(
+        "node1 and node2 are the same ('{}' == '{}'), but trace edges cannot "
+        "be removed by _remove_edges.".format(node1, node2))
+
+  node1_edges = node1.edges[:]
+  node2_edges = node2.edges[:]
+
+  nodes_set = set([node1, node2])
+  for edge in edges:
+    if edge.is_dangling():
+      raise ValueError("Attempted to remove dangling edge '{}'.".format(edge))
+    if set([edge.node1, edge.node2]) != nodes_set:
+      raise ValueError(
+          "Attempted to remove edges belonging to different node pairs: "
+          "'{}' != '{}'.".format(nodes_set, set([edge.node1, edge.node2])))
+
+  node1_axis_names = node1.axis_names
+  node2_axis_names = node2.axis_names
+
+  remaining_edges = []
+  for (i, edge) in enumerate(node1_edges):
+    if edge not in edges:  # NOTE: Makes the cost quadratic in # edges
+      edge.update_axis(
+          old_node=node1,
+          old_axis=i,
+          new_axis=len(remaining_edges),
+          new_node=new_node)
+      remaining_edges.append(edge)
+
+  for (i, edge) in enumerate(node2_edges):
+    if edge not in edges:
+      edge.update_axis(
+          old_node=node2,
+          old_axis=i,
+          new_axis=len(remaining_edges),
+          new_node=new_node)
+      remaining_edges.append(edge)
+
+  for (i, edge) in enumerate(remaining_edges):
+    new_node.add_edge(edge, i)
+
+  node1.fresh_edges(node1_axis_names)
+  node2.fresh_edges(node2_axis_names)
+  # pylint: disable=expression-not-assigned
+  [edge.disable() for edge in edges]  #disabled edges!
+
+
+def _contract_trace(edge: Edge, name: Optional[Text] = None) -> BaseNode:
+  """Contract a trace edge.
+  `edge` is disabled before returning.
+  Args:
+    edge: The edge name or object to contract next.
+    name: Name to give to the new node. If None, a name will automatically be
+      generated.
+
+  Returns:
+    The new node created after the contraction.
+
+  Raise:
+    ValueError: When edge is a dangling edge.
+  """
+  if edge.is_dangling():
+    raise ValueError("Attempted to contract dangling edge '{}'".format(edge))
+  if edge.node1 is not edge.node2:
+    raise ValueError("Can not take trace of edge '{}'. This edge connects to "
+                     "two different nodes: '{}' and '{}".format(
+                         edge, edge.node1, edge.node2))
+  backend = edge.node1.backend
+  axes = sorted([edge.axis1, edge.axis2])
+  dims = len(edge.node1.tensor.shape)
+  permutation = sorted(set(range(dims)) - set(axes)) + axes
+  new_tensor = backend.trace(
+      backend.transpose(edge.node1.tensor, perm=permutation))
+
+  new_node = Node(new_tensor, name=name, backend=backend.name)
+  _remove_trace_edge(edge, new_node)  #disables edge
+  return new_node
+
+
+def contract(edge: Edge,
+             name: Optional[Text] = None,
+             axis_names: Optional[List[Text]] = None) -> BaseNode:
+  """
+  Contract an edge connecting two nodes.
+  All edges of `node1` and `node2` are passed on to the new node,
+  and `node1` and `node2` get a new set of dangling edges.
+  `edge is disabled before returning.
+  Args:
+    edge: The edge contract next.
+    name: Name of the new node created.
+
+  Returns:
+    The new node created after the contraction.
+
+  Raises:
+    ValueError: When edge is a dangling edge or if it already has been
+      contracted.
+  """
+  if edge.is_dangling():
+    raise ValueError("Attempting to contract dangling edge")
+
+  for node in [edge.node1, edge.node2]:
+    if (node is not None) and (not hasattr(node, 'backend')):
+      raise TypeError('Node {} of type {} has no `backend`'.format(
+          node, type(node)))
+
+  if edge.node1.backend.name != edge.node2.backend.name:
+    raise ValueError("edge.node1 {} and edge.node2 {} have different backends "
+                     "{} and {}".format(edge.node1.name, edge.node2.name,
+                                        edge.node1.backend.name,
+                                        edge.node2.backend.name))
+
+  if edge.node1:
+    backend = edge.node1.backend
+  else:
+    raise ValueError("edge {} has no nodes. "
+                     "Cannot perfrom a contraction".format(edge.name))
+
+  backend = edge.node1.backend
+  if edge.node1 is edge.node2:
+    return _contract_trace(edge, name)
+  new_tensor = backend.tensordot(edge.node1.tensor, edge.node2.tensor,
+                                 [[edge.axis1], [edge.axis2]])
+  new_node = Node(
+      tensor=new_tensor, name=name, axis_names=axis_names, backend=backend.name)
+  # edge.node1 and edge.node2 get new edges in _remove_edges
+  _remove_edges(set([edge]), edge.node1, edge.node2, new_node)
+  return new_node
+
+
+def contract_copy_node(copy_node: CopyNode,
+                       name: Optional[Text] = None) -> BaseNode:
+  """Contract all edges incident on given copy node.
+
+  Args:
+    copy_node: Copy tensor node to be contracted.
+    name: Name of the new node created.
+
+  Returns:
+    New node representing contracted tensor.
+
+  Raises:
+    ValueError: If copy_node has dangling edge(s).
+  """
+  new_tensor = copy_node.compute_contracted_tensor()
+  new_node = Node(new_tensor, name, backend=copy_node.backend.name)
+
+  partners = copy_node.get_partners()
+  new_axis = 0
+  for partner in partners:
+    for edge in partner.edges:
+      if edge.node1 is copy_node or edge.node2 is copy_node:
+        continue
+      old_axis = edge.axis1 if edge.node1 is partner else edge.axis2
+      edge.update_axis(
+          old_node=partner,
+          old_axis=old_axis,
+          new_node=new_node,
+          new_axis=new_axis)
+      new_node.add_edge(edge, new_axis)
+      new_axis += 1
+  assert len(new_tensor.shape) == new_axis
+  copy_node.fresh_edges(copy_node.axis_names)
+  return new_node
+
+
+def contract_parallel(edge: Edge) -> BaseNode:
+  """Contract all edges parallel to this edge.
+
+    This method calls `contract_between` with the nodes connected by the edge.
+
+    Args:
+      edge: The edge to contract.
+
+    Returns:
+      The new node created after contraction.
+    """
+  if edge.is_dangling():
+    raise ValueError("Attempted to contract dangling edge: '{}'".format(edge))
+  return contract_between(edge.node1, edge.node2)
+
+
 def connect(edge1: Edge, edge2: Edge, name: Optional[Text] = None) -> Edge:
   for edge in [edge1, edge2]:
     if not edge.is_dangling():
@@ -1132,3 +1594,189 @@ def disconnect(edge, edge1_name: Optional[Text],
   and adding new dangling edges instead
   """
   return edge.disconnect(edge1_name, edge2_name)
+
+
+def contract_between(
+    node1: BaseNode,
+    node2: BaseNode,
+    name: Optional[Text] = None,
+    allow_outer_product: bool = False,
+    output_edge_order: Optional[Sequence[Edge]] = None,
+    axis_names: Optional[List[Text]] = None,
+) -> BaseNode:
+  """Contract all of the edges between the two given nodes.
+
+  Args:
+    node1: The first node.
+    node2: The second node.
+    name: Name to give to the new node created.
+    allow_outer_product: Optional boolean. If two nodes do not share any edges
+      and `allow_outer_product` is set to `True`, then we return the outer
+      product of the two nodes. Else, we raise a `ValueError`.
+    output_edge_order: Optional sequence of Edges. When not `None`, must
+      contain all edges belonging to, but not shared by `node1` and `node2`.
+      The axes of the new node will be permuted (if necessary) to match this
+      ordering of Edges.
+    axis_names: An optional list of names for the axis of the new node
+  Returns:
+    The new node created.
+
+  Raises:
+    ValueError: If no edges are found between node1 and node2 and
+      `allow_outer_product` is set to `False`.
+  """
+  for node in [node1, node2]:
+    if not hasattr(node, 'backend'):
+      raise TypeError('Node {} of type {} has no `backend`'.format(
+          node, type(node)))
+
+  if node1.backend.name != node2.backend.name:
+    raise ValueError("node {} and node {} have different backends "
+                     "{} and {}.".format(node1.name, node2.name,
+                                         node1.backend.name,
+                                         node2.backend.name))
+
+  backend = node1.backend
+  # Trace edges cannot be contracted using tensordot.
+  if node1 is node2:
+    flat_edge = flatten_edges_between(node1, node2)
+    if not flat_edge:
+      raise ValueError("No trace edges found on contraction of edges between "
+                       "node '{}' and itself.".format(node1))
+    return contract(flat_edge, name)
+
+  shared_edges = get_shared_edges(node1, node2)
+  if not shared_edges:
+    if allow_outer_product:
+      return outer_product(node1, node2, name=name, axis_names=axis_names)
+    raise ValueError("No edges found between nodes '{}' and '{}' "
+                     "and allow_outer_product=False.".format(node1, node2))
+
+  # Collect the axis of each node corresponding to each edge, in order.
+  # This specifies the contraction for tensordot.
+  # NOTE: The ordering of node references in each contraction edge is ignored.
+  axes1 = []
+  axes2 = []
+  for edge in shared_edges:
+    if edge.node1 is node1:
+      axes1.append(edge.axis1)
+      axes2.append(edge.axis2)
+    else:
+      axes1.append(edge.axis2)
+      axes2.append(edge.axis1)
+
+  if output_edge_order:
+    # Determine heuristically if output transposition can be minimized by
+    # flipping the arguments to tensordot.
+    node1_output_axes = []
+    node2_output_axes = []
+    for (i, edge) in enumerate(output_edge_order):
+      if edge in shared_edges:
+        raise ValueError(
+            "Edge '{}' in output_edge_order is shared by the nodes to be "
+            "contracted: '{}' and '{}'.".format(edge, node1, node2))
+      edge_nodes = set(edge.get_nodes())
+      if node1 in edge_nodes:
+        node1_output_axes.append(i)
+      elif node2 in edge_nodes:
+        node2_output_axes.append(i)
+      else:
+        raise ValueError(
+            "Edge '{}' in output_edge_order is not connected to node '{}' or "
+            "node '{}'".format(edge, node1, node2))
+    if np.mean(node1_output_axes) > np.mean(node2_output_axes):
+      node1, node2 = node2, node1
+      axes1, axes2 = axes2, axes1
+
+  new_tensor = backend.tensordot(node1.tensor, node2.tensor, [axes1, axes2])
+  new_node = Node(
+      tensor=new_tensor, name=name, axis_names=axis_names, backend=backend.name)
+  # node1 and node2 get new edges in _remove_edges
+  _remove_edges(shared_edges, node1, node2, new_node)
+  if output_edge_order:
+    new_node = new_node.reorder_edges(list(output_edge_order))
+  return new_node
+
+
+def outer_product_final_nodes(nodes: Union[List[BaseNode], Set[BaseNode]],
+                              edge_order: List[Edge]) -> BaseNode:
+  """Get the outer product of `nodes`
+
+  For example, if there are 3 nodes remaining in `nodes` with 
+  shapes :math:`(2, 3)`, :math:`(4, 5, 6)`, and :math:`(7)`
+  respectively, the newly returned node will have shape 
+  :math:`(2, 3, 4, 5, 6, 7)`.
+
+  Args:
+    nodes: A collection of nodes.
+    edge_order: Edge order for the final node.
+
+  Returns:
+    The outer product of the remaining nodes.
+
+  Raises:
+    ValueError: If any of the remaining nodes are not fully contracted.
+  """
+  nodes = list(nodes)
+  for node in nodes:
+    if node.has_nondangling_edge():
+      raise ValueError("Node '{}' has a non-dangling edge remaining.")
+  final_node = nodes[0]
+  for node in nodes[1:]:
+    final_node = outer_product(final_node, node)
+  return final_node.reorder_edges(edge_order)
+
+
+def outer_product(node1: BaseNode,
+                  node2: BaseNode,
+                  name: Optional[Text] = None,
+                  axis_names: Optional[List[Text]] = None) -> BaseNode:
+  """Calculates an outer product of the two nodes.
+
+  This causes the nodes to combine their edges and axes, so the shapes are
+  combined. For example, if `a` had a shape (2, 3) and `b` had a shape
+  (4, 5, 6), then the node `net.outer_product(a, b) will have shape
+  (2, 3, 4, 5, 6). All edges of `node1` and `node2` are passed on to
+  the new node, and `node1` and `node2` get a new set of dangling edges.
+
+  Args:
+    node1: The first node. The axes on this node will be on the left side of
+      the new node.
+    node2: The second node. The axes on this node will be on the right side of
+      the new node.
+    name: Optional name to give the new node created.
+    axis_names: An optional list of names for the axis of the new node
+  Returns:
+    A new node. Its shape will be node1.shape + node2.shape
+  Raises:
+    TypeError: If `node1` and `node2` have wrong types.
+  """
+  for node in [node1, node2]:
+    if not hasattr(node, 'backend'):
+      raise TypeError('Node {} of type {} has no `backend`'.format(
+          node, type(node)))
+
+  if node1.backend.name != node2.backend.name:
+    raise ValueError("node {}  and node {} have different backends. "
+                     "Cannot perform outer product".format(node1, node2))
+
+  backend = node1.backend
+  new_tensor = backend.outer_product(node1.tensor, node2.tensor)
+  node1_axis_names = node1.axis_names
+  node2_axis_names = node2.axis_names
+  new_node = Node(
+      tensor=new_tensor, name=name, axis_names=axis_names, backend=backend.name)
+  additional_axes = len(node1.tensor.shape)
+
+  for i, edge in enumerate(node1.edges):
+    edge.update_axis(i, node1, i, new_node)
+  for i, edge in enumerate(node2.edges):
+    edge.update_axis(i, node2, i + additional_axes, new_node)
+
+  for i, edge in enumerate(node1.edges + node2.edges):
+    new_node.add_edge(edge, i, True)
+
+  node1.fresh_edges(node1_axis_names)
+  node2.fresh_edges(node2_axis_names)
+
+  return new_node
